Break up emitter module
