+++ codespan-reporting/src/emitter/views/mod.rs
@@ -1,0 +1,21 @@
+mod border_left;
+mod border_top;
+mod border_top_left;
+mod gutter;
+mod header;
+mod locus;
+mod new_line;
+mod note;
+mod note_bullet;
+mod source_snippet;
+
+pub use self::border_left::*;
+pub use self::border_top::*;
+pub use self::border_top_left::*;
+pub use self::gutter::*;
+pub use self::header::*;
+pub use self::locus::*;
+pub use self::new_line::*;
+pub use self::note::*;
+pub use self::note_bullet::*;
+pub use self::source_snippet::*;
