+++ codespan-reporting/src/emitter/views/new_line.rs
@@ -1,0 +1,17 @@
+use std::io;
+use termcolor::WriteColor;
+
+use crate::emitter::Config;
+
+/// A new line.
+pub struct NewLine {}
+
+impl<'a> NewLine {
+    pub fn new() -> NewLine {
+        NewLine {}
+    }
+
+    pub fn emit(&self, writer: &mut impl WriteColor, _config: &Config) -> io::Result<()> {
+        write!(writer, "\n")
+    }
+}
